pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "yacht_osint"
version = "0.1.0"
description = "Automated OSINT scraper for super-yachts"
authors = [{name = "Tenderworks Prodigy"}]
readme = "README.md"
requires-python = ">=3.11"

# specify dependencies from requirements.txt
dependencies = [
    "requests==2.31.0",
    "cloudscraper==1.2.71",
    "feedparser==6.0.11",
    "beautifulsoup4==4.12.3",
    "lxml==4.9.3",
    "trafilatura==1.6.1",
    "waybackpy==3.0.6",
    "pandas==2.1.4",
    "duckdb==1.1.2",
    "pyarrow==14.0.2",
    "numpy==1.26.4",
    "scikit-learn==1.4.2",
    "pydantic==1.10.12",
    "jsonschema==4.19.2",
    "jinja2==3.1.2",
    "tqdm==4.66.2",
    "gspread==6.1.2",
    "oauth2client==4.1.3",
    "pydrive2==1.15.1",
    "pandera==0.17.2",
    "tenacity==8.2.3",
    "python-dotenv==1.0.0",
    "feedfinder2==0.0.4",
    "tldextract==3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "ruff==0.12.5",
    "black",
    "pytest-cov",
    "vcrpy==5.1.0",
]

[project.scripts]
yacht-osint = "src.cli:run"
